tflite-requirements.txt
pillow
# Requirements for running inference on embedded devices
tflite-runtime==2.11.0
